Add vscode config to prefer node_modules TS
This config (optionally) hints to the Typescript extension that the project's
Typescript version (which is often bleeding edge/"RC") should be used. As this
is a HelloWorld mod project, my hope is that it might be useful for
mitigating yet another confusing roadblock for would be modders in the event
of some TS incompatibility.